#8 - ALPN API should allow throwing IOException or SSLException.
See https://github.com/jetty-project/jetty-alpn/issues/8.

Turns out that the OpenJDK SSL implementation does different things
in SSLSocket versus SSLEngine when an exception is thrown: in the
first case it wraps it into an SSLException, in the second it does not.

By allowing the API to explicitly throw an SSLException, the behavior
is now more consistent.